# src/api/swagger/portal-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: wicked.haufe.io Portal API
  description: The API which the API Portal uses to store and retrieve information
 
# Under securityDefinitions we declare which security schemes can be used.
# These definitions do not affect to the operations, but can be applied
# using the 'security' key at the global level or the operation level.
# In this sample, we'll see how to apply the security definition globally,
# and override it at the operation level.
#
# Note that for Access Code Flow, you must provide the Token URL.
securityDefinitions:
  # For personalized access to data
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/auth/local/portal-api/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/auth/local/portal-api/oauth2/authorize
    flow: accessCode
    scopes:
      read_plans: Read plans
      read_groups: Read user group definitions
      read_content: Read custom/markdown/HTML content
      health: Read out health information
      read_templates: Read templates
      read_auth_servers: Read authorization servers
      read_apis: Read API definitions
      read_users: Read user information
      write_users: Write user information
      read_applications: Read application data
      write_applications: Write application data
      read_subscriptions: Read subscriptions
      write_subscriptions: Write/edit/delete subscriptions
      read_approvals: Read pending approvals
      read_auditlog: Read audit log
      read_verifications: Read email and password verifications
      write_verifications: Create or delete verifications
      read_registrations: Read user registrations
      write_registrations: Write user registrations
      read_grants: Read user API grants
      write_grants: Write user API grants
      webhooks: Access and change webhooks
      restart_api: Restart the API components
      access_tokens: Administer access tokens
  
  implicit:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/portal-api/v1/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/portal-auth/api/portal-api
...
produces:
  - application/hal+json
host: 'localhost:3001'
basePath: /portal-api/v1
...
paths:
...
  '/accesstokens':
    post:
      security:
      - accessCode: ['access_tokens']
      - implicit: ['access_tokens']
      - password: ['access_tokens']
      tags:
      - accesstokens
      description: Add an access token registration
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/accessTokenType'
      responses:
        '201':
          description: Success, created
        '403':
          description: Forbidden
        '400':
          description: Bad request
    get:
      security:
      - accessCode: ['access_tokens']
      - implicit: ['access_tokens']
      - password: ['access_tokens']
      tags:
      - accesstokens
      description: Get access tokens by access token, refresh token or authenticated user id
      parameters:
      - name: access_token
        in: query
        required: false
        type: string
      - name: refresh_token
        in: query
        required: false
        type: string
      - name: authenticated_userid
        in: query
        required: false
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/accessTokenCollection'
        '403':
          description: Forbidden

    delete:
      security:
      - accessCode: ['access_tokens']
      - implicit: ['access_tokens']
      - password: ['access_tokens']
      tags:
      - accesstokens
      description: Delete access tokens by access token, refresh token or authenticated user id
      parameters:
      - name: access_token
        in: query
        required: false
        type: string
      - name: refresh_token
        in: query
        required: false
        type: string
      - name: authenticated_userid
        in: query
        required: false
        type: string
      responses:
        '204':
          description: Success
        '403':
          description: Forbidden
definitions:
...
  accessTokenType:
    type: object
    properties:
      access_token:
        type: string
        example: 'abc76afF898a8ADopiaHDS987s79dsd'
      expires:
        type: number
        example: 1628267233252
      refresh_token:
        type: string
        example: 'JK8a6sf98ljDFDFkjlhgjASDf813912'
      expires_refresh:
        type: number
        example: 1628267233252
      authenticated_userid:
        type: string
        example: 'sub=e8ac050b4019405218fa5d51521e0697768fe63d;namespace=wicked'
      users_id:
        type: string
        example: 'e8ac050b4019405218fa5d51521e0697768fe63d'
      scope:
        type: string
        example: 'access_tokens read_apis write_users'
      application_id:
        type: string
        example: 'some-application'
      api_id:
        type: string
        example: 'petstore-oauth'
      plan_id:
        type: string
        example: 'basic'
      subscription_id:
        type: string
        example: '940521801521e0697768fe63dfa5d51e8ac050b4'
  accessTokenCollection:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/accessTokenType'
      count:
        type: number
        example: 1
      count_cached:
        type: boolean
        example: 'false'
      offset:
        type: number
        example: 0
      limit:
        type: number
        example: 1